Support startCheckout with a Fragment
https://github.com/Affirm/affirm-merchant-sdk-android/issues/80
