+++ affirm/src/main/java/com/affirm/android/AffirmActivity.java
@@ -1,18 +1,33 @@
 package com.affirm.android;
 
+import android.app.Activity;
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.view.ViewGroup;
 import android.webkit.WebView;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.fragment.app.Fragment;
 
 abstract class AffirmActivity extends AppCompatActivity implements AffirmWebChromeClient.Callbacks {
 
     ViewGroup container;
     WebView webView;
     View progressIndicator;
+
+    static void startForResult(@NonNull Activity originalActivity,
+                               @Nullable Fragment fragment,
+                               @NonNull Intent intent,
+                               int requestCode) {
+        if (fragment != null) {
+            fragment.startActivityForResult(intent, requestCode);
+        } else {
+            originalActivity.startActivityForResult(intent, requestCode);
+        }
+    }
 
     abstract void initViews();
