+++ samples-java/src/main/res/layout/activity_main.xml
@@ -1,148 +1,5 @@
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
+<fragment xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/main_fragment"
+    android:name="com.affirm.samples.MainFragment"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <RelativeLayout xmlns:tools="http://schemas.android.com/tools"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content">
-
-        <TextView
-            android:id="@+id/title"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignTop="@+id/imageView"
-            android:layout_marginStart="20dp"
-            android:layout_marginLeft="20dp"
-            android:layout_toEndOf="@+id/imageView"
-            android:layout_toRightOf="@+id/imageView"
-            android:text="Great Wheel Deal" />
-
-        <TextView
-            android:id="@+id/promotionTextView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/price"
-            android:layout_margin="8dp"
-            android:gravity="center"
-            android:textColor="@android:color/black"
-            android:textSize="18sp"
-            tools:text="Pay as low as $69/month with Affirm" />
-
-        <com.affirm.android.PromotionWebView
-            android:id="@+id/htmlPromotionWebView"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/promotionTextView"
-            android:layout_marginTop="20dp"
-            android:gravity="center" />
-
-        <com.affirm.android.AffirmPromotionButton
-            android:id="@+id/promo"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/htmlPromotionWebView"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="20dp"
-            app:affirmColor="AffirmColorTypeBlue"
-            app:affirmLogoType="AffirmDisplayTypeLogo"
-            app:affirmTextColor="@android:color/black"
-            app:affirmTextFont="@font/apercu_bold"
-            app:affirmTextSize="@dimen/affirm_promotion_size"
-            app:htmlStyling="false" />
-
-        <FrameLayout
-            android:id="@+id/promo_container"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/promo"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="20dp" />
-
-        <EditText
-            android:id="@+id/cass"
-            android:layout_marginHorizontal="60dp"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:singleLine="true"
-            android:layout_below="@id/promo_container"
-            android:hint="cass"/>
-
-        <Button
-            android:id="@+id/checkout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/cass"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="16dp"
-            android:text="@string/checkout" />
-
-        <Button
-            android:id="@+id/vcnCheckout"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/checkout"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="16dp"
-            android:text="@string/vcn_checkout" />
-
-        <Button
-            android:id="@+id/siteModalButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/vcnCheckout"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="16dp"
-            android:text="@string/launch_site_modal" />
-
-        <Button
-            android:id="@+id/productModalButton"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@id/siteModalButton"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="16dp"
-            android:text="@string/product_site_modal" />
-
-        <Button
-            android:id="@+id/trackOrderConfirmed"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/productModalButton"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="16dp"
-            android:text="@string/track_order_confirmed" />
-
-        <Button
-            android:id="@+id/clearCookies"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/trackOrderConfirmed"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="16dp"
-            android:text="@string/clear_cookies" />
-
-        <ImageView
-            android:id="@+id/imageView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignParentStart="true"
-            android:layout_alignParentLeft="true"
-            android:layout_alignParentTop="true"
-            android:layout_marginStart="27dp"
-            android:layout_marginLeft="27dp"
-            android:layout_marginTop="19dp"
-            app:srcCompat="@mipmap/ic_launcher" />
-
-        <TextView
-            android:id="@+id/price"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_alignStart="@+id/title"
-            android:layout_alignLeft="@+id/title"
-            android:layout_alignBottom="@+id/imageView"
-            android:ems="10"
-            android:textSize="18sp"
-            android:textStyle="bold" />
-    </RelativeLayout>
-</ScrollView>
+    android:layout_height="match_parent" />
